Move API type tests to integration suite
