+++ pkg/api/kuberhealthycheck_types_integration_test.go
@@ -1,0 +1,66 @@
+//go:build integration
+// +build integration
+
+package api
+
+import (
+	"context"
+	"encoding/json"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+)
+
+// TestGetCheckSetsCreationTimestamp ensures GetCheck populates metadata.CreationTimestamp when missing.
+func TestGetCheckSetsCreationTimestamp(t *testing.T) {
+	scheme := runtime.NewScheme()
+	require.NoError(t, AddToScheme(scheme))
+
+	check := &KuberhealthyCheck{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "test",
+			Namespace: "default",
+		},
+	}
+
+	cl := fake.NewClientBuilder().WithScheme(scheme).WithObjects(check).Build()
+	nn := types.NamespacedName{Name: "test", Namespace: "default"}
+
+	out, err := GetCheck(context.Background(), cl, nn)
+	require.NoError(t, err)
+	require.False(t, out.CreationTimestamp.IsZero(), "creation timestamp must be set")
+}
+
+// TestSpecDoesNotExposeCreationTimestamp ensures creationTimestamp is not serialized in pod metadata.
+func TestSpecDoesNotExposeCreationTimestamp(t *testing.T) {
+	check := &KuberhealthyCheck{
+		Spec: KuberhealthyCheckSpec{
+			PodSpec: CheckPodSpec{
+				Metadata: &CheckPodMetadata{Labels: map[string]string{"foo": "bar"}},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{{
+						Name:  "test",
+						Image: "busybox",
+					}},
+				},
+			},
+		},
+	}
+
+	raw, err := json.Marshal(check)
+	require.NoError(t, err)
+
+	var data map[string]any
+	require.NoError(t, json.Unmarshal(raw, &data))
+
+	spec := data["spec"].(map[string]any)
+	podSpec := spec["podSpec"].(map[string]any)
+	metadata := podSpec["metadata"].(map[string]any)
+	_, found := metadata["creationTimestamp"]
+	require.False(t, found, "creationTimestamp must be absent in spec.podSpec.metadata")
+}
